--- crates/cranelift-shared/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "wasmtime-cranelift-shared"
-version.workspace = true
-authors.workspace = true
-description = "Base-level integration with Wasmtime and Cranelift"
-license = "Apache-2.0 WITH LLVM-exception"
-repository = "https://github.com/bytecodealliance/wasmtime"
-documentation = "https://docs.rs/wasmtime-cranelift-shared/"
-edition.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-wasmtime-environ = { workspace = true }
-cranelift-codegen = { workspace = true }
-cranelift-control = { workspace = true }
-cranelift-native = { workspace = true }
-target-lexicon = { workspace = true }
-gimli = { workspace = true }
-object = { workspace = true }
