+++ crates/cranelift/src/compiled_function.rs
