+++ crates/cranelift/src/compiler/component.rs
@@ -1,13 +1,15 @@
 //! Compilation support for the component model.
 
-use crate::compiler::{Compiler, NativeRet};
+use crate::{
+    compiler::{Compiler, NativeRet},
+    ALWAYS_TRAP_CODE, CANNOT_ENTER_CODE,
+};
 use anyhow::Result;
 use cranelift_codegen::ir::{self, InstBuilder, MemFlags};
 use cranelift_codegen::isa::{CallConv, TargetIsa};
 use cranelift_frontend::FunctionBuilder;
 use cranelift_wasm::ModuleInternedTypeIndex;
 use std::any::Any;
-use wasmtime_cranelift_shared::{ALWAYS_TRAP_CODE, CANNOT_ENTER_CODE};
 use wasmtime_environ::component::*;
 use wasmtime_environ::{PtrSize, WasmValType};
