+++ crates/cranelift/src/debug/transform/expression.rs
@@ -1115,8 +1115,8 @@ mod tests {
     }
 
     fn create_mock_address_transform() -> AddressTransform {
+        use crate::FunctionAddressMap;
         use cranelift_entity::PrimaryMap;
-        use wasmtime_cranelift_shared::FunctionAddressMap;
         use wasmtime_environ::InstructionAddressMap;
         use wasmtime_environ::WasmFileInfo;
